Moved files in Engine & replaced dirlist with SPL
