+++ tests/Integration/inc/Engine/Optimization/CSS/CriticalCSS/cleanCriticalCss.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace WP_Rocket\Tests\Integration\inc\classes\optimization\CSS\Critical_CSS;
+namespace WP_Rocket\Tests\Integration\inc\Engine\Optimization\CSS\CriticalCSS;
 
 use Mockery;
 use Brain\Monkey\Functions;
 use WP_Rocket\Tests\Integration\FilesystemTestCase;
 
 /**
- * @covers \WP_Rocket\Optimization\CSS\Critical_CSS::clean_critical_css
+ * @covers \WP_Rocket\Engine\Optimization\CSS\CriticalCSS::clean_critical_css
  * @group  AdminOnly
  * @group  CPCSS
  */
 class Test_CleanCriticalCSS extends FilesystemTestCase {
-	protected $path_to_test_data = '/inc/classes/optimization/CSS/Critical_CSS/cleanCriticalCss.php';
+	protected $path_to_test_data = '/inc/Engine/Optimization/CSS/CriticalCSS/cleanCriticalCss.php';
 
 	private $deleted_files = [
 		'home.css',
